Move search logic out of controller
